Cleanup Vertex + Chat
